Login de Cliente em processo

+++ src/app/controllers/formaRecebimentoController.js
@@ -7,12 +7,16 @@ const mysql = require('../../database/index').pool
 router.get('/pix', (req, res, next) => {
 
     mysql.getConnection((error, conn) => {
-        conn.query(`SELECT * FROM tblPixArtista`, function(error, rows, fields) {
+        if (error) { return res.status(500).send({ error: error }) } 
+        conn.query(`SELECT * FROM tblPixArtista`, 
+        
+        (error, results, fields) => {
+            conn.release()
 
             if (error) { return res.status(500).send({ error: error }) } 
 
             res.status(200).send({
-                    pix: rows
+                    pix: results
             })
         })
     })
@@ -25,19 +29,16 @@ router.get('/pix/:artistaId', (req, res, next) => {
     const id = req.params.artistaId
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(`SELECT * FROM tblPixArtista WHERE idArtista = ?`, [id], 
-        function(err, rows, fields) {
+        (error, results, fields) => {
 
-            if (error) {
-                console.log(error);
-                return res.status(500).send({
-                    error: error,
-                    response: null
-                })
-            } 
+            conn.release()
+
+            if (error) { return res.status(500).send({ error: error }) } 
 
             res.status(200).send({
-                    pixDeArtista: rows
+                    pixDeArtista: results
             })
         })
     })
@@ -53,6 +54,7 @@ router.post('/pix/:artistaId', (req, res, next) => {
     } = req.body
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(
             `INSERT INTO tblPixArtista(tipoChave, chave, idArtista) 
                 VALUES(?,?,?)`,
@@ -63,15 +65,32 @@ router.post('/pix/:artistaId', (req, res, next) => {
                 
                 if (error) { return res.status(500).send({ error: error }) } 
 
-                res.status(201).send({
+                const response = {
+                    
+                    tipoChave: req.body.tipoChave,
+                    chave: req.body.chave, 
                     mensagem: 'Pix de Artista cadastrado com sucesso'
-                })
+            
+                }
 
-                conn.query(
-                   `DELETE FROM tblContaBancariaArtista WHERE idArtista = ?`, [id],
-                   conn.release()
+                conn.query('SELECT * FROM tblContaBancariaArtista WHERE idArtista = ?', [id], 
+                    (error, results) => {
+                        conn.release()
+                        if (error) { return res.status(500).send({ error: error }) } 
+                        if (results.length > 0){
+                            conn.query(
+                                `DELETE FROM tblContaBancariaArtista WHERE idArtista = ?`, [id],
+                                (error, results) => {
+                                    conn.release()
+                                    if (error) { return res.status(500).send({ error: error }) } 
+                                }
+                            )
+                        }
+                    }
                 )
-                
+
+                res.status(201).send(response)
+
             }
         )
     })
@@ -88,6 +107,7 @@ router.patch('/pix/:artistaId', (req, res, next) => {
     } = req.body
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(
             `UPDATE tblPixArtista SET tipoChave = ?, chave = ? WHERE idArtista = ?`,
             [tipoChave,chave,id],
@@ -97,11 +117,15 @@ router.patch('/pix/:artistaId', (req, res, next) => {
                 
                 if (error) { return res.status(500).send({ error: error }) } 
 
-                res.status(201).send({
-                    mensagem: 'Pix de Artista atualizado com sucesso'
-                })
+                const response = {
 
-            
+                            tipoChave: req.body.tipoChave,
+                            chave: req.body.chave, 
+                            mensagem: 'Pix de Artista atualizado com sucesso'
+                     
+                }
+
+                res.status(201).send(response)
 
             }
         )
@@ -113,14 +137,15 @@ router.patch('/pix/:artistaId', (req, res, next) => {
 router.get('/contaBancaria', (req, res, next) => {
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(`SELECT * FROM tblContaBancariaArtista`, 
         
-        function(error, rows, fields) {
+        (error, results, fields) => {
 
             if (error) { return res.status(500).send({ error: error }) } 
 
             res.status(200).send({
-                    contasBancarias: rows
+                    contasBancarias: results
             })
         })
     })
@@ -132,14 +157,15 @@ router.get('/contaBancaria/:artistaId', (req, res, next) => {
     const id = req.params.artistaId
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(`SELECT * FROM tblContaBancariaArtista WHERE idArtista = ?`, [id],
 
-        function(err, rows, fields) {
+        (error, results, fields) => {
 
             if (error) { return res.status(500).send({ error: error }) } 
 
             res.status(200).send({
-                    contaBancariaDeArtista: rows
+                    contaBancariaDeArtista: results
             })
         })
     })
@@ -156,6 +182,7 @@ router.post('/contaBancaria/:artistaId', (req, res, next) => {
     } = req.body
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(
             `INSERT INTO tblContaBancariaArtista(tipoConta, banco, titular, cpfTitular, 
                 agencia, digito, conta, digitoVerificador, idArtista) 
@@ -167,14 +194,37 @@ router.post('/contaBancaria/:artistaId', (req, res, next) => {
                 
                 if (error) { return res.status(500).send({ error: error }) } 
 
-                res.status(201).send({
-                    mensagem: 'Conta Bancária de Artista cadastrado com sucesso'
-                })
+                const response = {
+                    
+                        tipoConta: req.body.tipoConta,
+                        banco: req.body.banco,
+                        titular: req.body.titular,
+                        cpfTitular: req.body.cpfTitular,
+                        agencia: req.body.agencia,
+                        digito: req.body.digito,
+                        conta: req.body.conta,
+                        digitoVerificador: req.body.digitoVerificador,
+                        mensagem: 'Conta Bancária de Artista cadastrado com sucesso' 
+                     
+                }
 
-                conn.query(
-                    `DELETE FROM tblPixArtista WHERE idArtista = ?`, [id],
-                    conn.release()
-                 )
+                conn.query('SELECT * FROM tblPixArtista WHERE idArtista = ?', [id], 
+                    (error, results) => {
+                        conn.release()
+                        if (error) { return res.status(500).send({ error: error }) } 
+                        if (results.length > 0){
+                            conn.query(
+                                `DELETE FROM tblPixArtista WHERE idArtista = ?`, [id],
+                                (error, results) => {
+                                    conn.release()
+                                    if (error) { return res.status(500).send({ error: error }) } 
+                                }
+                            )
+                        }
+                    }
+                )
+
+                res.status(201).send(response)
                     
 
             }
@@ -193,6 +243,7 @@ router.patch('/contaBancaria/:artistaId', (req, res, next) => {
     } = req.body
 
     mysql.getConnection((error, conn) => {
+        if (error) { return res.status(500).send({ error: error }) } 
         conn.query(
             `UPDATE tblContaBancariaArtista SET tipoConta = ?, banco = ?, 
                                                 titular = ?, cpfTitular = ?, 
@@ -206,11 +257,21 @@ router.patch('/contaBancaria/:artistaId', (req, res, next) => {
                 
                 if (error) { return res.status(500).send({ error: error }) } 
 
-                res.status(201).send({
+                const response = {
+                   
+                    tipoConta: req.body.tipoConta,
+                    banco: req.body.banco,
+                    titular: req.body.titular,
+                    cpfTitular: req.body.cpfTitular,
+                    agencia: req.body.agencia,
+                    digito: req.body.digito,
+                    conta: req.body.conta,
+                    digitoVerificador: req.body.digitoVerificador,
                     mensagem: 'Conta Bancária de Artista atualizada com sucesso'
-                })
+                     
+                }
 
- 
+                res.status(201).send(response)
 
             }
         )
